+++ ChangeLog.md
@@ -1,12 +1,16 @@
 # Revision history for cesds-records
 
-## 0.1.3.0  -- 2016-06-24
+## 0.1.3.1  -- 2016-06-25
 
-* Rewrote JSON instances to conform to CESDS specification
+* Tested `CESDS.Types`
+
+## 0.1.3.0  -- 2016-06-25
+
+* Rewrote `FromJSON` and `ToJSON` instances to conform to CESDS specification.
 
 ## 0.1.2.0  -- 2016-06-24
 
-* Added orphan Read, FromJSON, and ToJSON instances for URI
+* Added orphan `Read`, `FromJSON`, and `ToJSON` instances for `URI`.
 
 ## 0.1.1.0  -- 2016-06-23